Fixed requirements.txt deletion bug on app startup.

+++ UVR.py
@@ -2273,7 +2273,7 @@ class MainWindow(TkinterDnD.Tk if is_dnd_compatible else tk.Tk):
             
             for dir in DIRECTORIES:
                 for temp_file in os.listdir(dir):
-                    if temp_file.endswith(EXTENSIONS):
+                    if temp_file.endswith(EXTENSIONS) and temp_file != 'requirements.txt': # HACK: Exclude requirements.txt to prevent accidental deletion
                         if os.path.isfile(os.path.join(dir, temp_file)):
                             os.remove(os.path.join(dir, temp_file))
         except Exception as e:
